#1767 cleanup
