+++ view/statistics-flow-roles.js
@@ -15,14 +15,17 @@ var _                 = require('mano').i18n.bind('View: Statistics')
   , selectCertificate = require('./components/filter-bar/select-certificate')
   , selectPeriodMode  = require('./components/filter-bar/select-period-mode')
   , itemsPerPage      = require('../conf/objects-list-items-per-page')
-  , isNaturalNumber   = require('es5-ext/number/is-natural')
   , serviceQuery      = require('../apps-common/query-conf/service')
   , certificateQuery  = require('../apps-common/query-conf/certificate')
   , stepStatusQuery   = require('../apps-common/query-conf/processing-step-status')
+  , pageQuery         = require('../utils/query/date-constrained-page')
   , copyDbDate        = require('../utils/copy-db-date')
   , queryServer       = require('./utils/statistics-flow-query-server')
   , processingSteps   = require('../processing-steps-meta')
-  , getStepLabelByShortPath = require('../utils/get-step-label-by-short-path');
+  , getStepLabelByShortPath = require('../utils/get-step-label-by-short-path')
+  , incrementDateByTimeUnit = require('../utils/increment-date-by-time-unit')
+  , floorToTimeUnit         = require('../utils/floor-to-time-unit')
+  , calculateDurationByMode = require('../utils/calculate-duration-by-mode');
 
 exports._parent        = require('./statistics-flow');
 exports._customFilters = Function.prototype;
@@ -62,44 +65,36 @@ var buildResultRow = function (rowData, queryCertificate, queryStatus) {
 		resultRow[stepKey] = 0;
 	});
 	Object.keys(rowData).forEach(function (stepShortPath) {
-		Object.keys(rowData[stepShortPath]).forEach(function (status) {
-			if (status !== queryStatus) return;
+		if (rowData[stepShortPath][queryStatus] == null) {
+			resultRow[stepShortPath] = null;
+			return;
+		}
 
-			reducedRowData = rowData[stepShortPath][status];
-			if (queryCertificate) {
-				reducedRowData = reducedRowData.certificate[queryCertificate];
-			} else {
-				reducedRowData = reducedRowData.businessProcess;
-			}
-			resultRow[stepShortPath] = reducedRowData;
-		});
+		reducedRowData = rowData[stepShortPath][queryStatus];
+		if (queryCertificate) {
+			reducedRowData = reducedRowData.certificate[queryCertificate];
+		} else {
+			reducedRowData = reducedRowData.businessProcess;
+		}
+		resultRow[stepShortPath] = reducedRowData;
 	});
 	return resultRow;
 };
 
-var incrementDateByTimeUnit = function (date, mode) {
-	switch (mode) {
-	case 'weekly':
-		date.setUTCDate(date.getUTCDate() + 7);
-		break;
-	case 'monthly':
-		date.setUTCMonth(date.getUTCMonth() + 1);
-		break;
-	case 'yearly':
-		date.setUTCFullYear(date.getUTCFullYear() + 1);
-		break;
-	default:
-		date.setUTCDate(date.getUTCDate() + 1);
-		break;
-	}
+var fillEmpty = function (rowData) {
+	Object.keys(rowData).forEach(function (stepShortPath) {
+		if (!rowData[stepShortPath]) {
+			rowData[stepShortPath] = _("N/A");
+		}
+	});
 
-	return date;
+	return rowData;
 };
 
 var buildFilteredResult = function (data, key, service, certificate, status) {
 	var resultRow, finalResult = {};
 	if (service) {
-		return buildResultRow(data[key][service].processingStep, certificate, status);
+		return fillEmpty(buildResultRow(data[key][service].processingStep, certificate, status));
 	}
 
 	Object.keys(data[key]).forEach(function (serviceKey) {
@@ -113,32 +108,7 @@ var buildFilteredResult = function (data, key, service, certificate, status) {
 		});
 	});
 
-	return finalResult;
-};
-
-var floorToUnit = function (date, mode) {
-	switch (mode) {
-	case 'weekly':
-		var dayOfWeek = date.getUTCDay(), modifier;
-		if (dayOfWeek === 0) {
-			modifier = -6;
-		} else {
-			modifier = -(dayOfWeek - 1);
-		}
-		date.setUTCDate(date.getUTCDate() + modifier);
-		break;
-	case 'monthly':
-		date.setUTCDate(1);
-		break;
-	case 'yearly':
-		date.setUTCDate(1);
-		date.setUTCMonth(0);
-		break;
-	default:
-		break;
-	}
-
-	return date;
+	return fillEmpty(finalResult);
 };
 
 var filterData = function (data, query) {
@@ -154,7 +124,7 @@ var filterData = function (data, query) {
 			return true;
 		}
 	});
-	floorToUnit(currentDate, mode.key);
+	floorToTimeUnit(currentDate, mode.key);
 	while (currentDate <= dateTo) {
 		key         = mode.getDisplayedKey(currentDate);
 		result[key] = buildFilteredResult(data, key, service, certificate, status);
@@ -164,44 +134,12 @@ var filterData = function (data, query) {
 	return result;
 };
 
-var calculateDurationByMode = function (dateFrom, dateTo, mode) {
-	var timeUnitsCount = 0, currentDate;
-	currentDate = copyDbDate(dateFrom);
-	if (!dateTo) dateTo = new db.Date();
-
-	floorToUnit(currentDate, mode);
-	while (currentDate <= dateTo) {
-		incrementDateByTimeUnit(currentDate, mode);
-		timeUnitsCount++;
-	}
-	return timeUnitsCount;
-};
-
 exports['statistics-main'] = function () {
 	var queryHandler, data = new ObservableValue({})
 	  , pagination = new Pagination('/flow/by-role/'), handlerConf;
 
 	handlerConf = queryHandlerConf.slice(0);
-	handlerConf.push({
-		name: 'page',
-		ensure: function (value, resolvedQuery, query) {
-			var num, dateFrom, dateTo, mode, durationInTimeUnits, resolvedValue;
-			resolvedValue = value;
-			if (resolvedValue == null) resolvedValue = '1';
-			if (isNaN(resolvedValue)) throw new Error("Unrecognized page value " + JSON.stringify(value));
-			num = Number(resolvedValue);
-			if (!isNaturalNumber(num)) throw new Error("Unreconized page value " + JSON.stringify(value));
-			if (num < 1) throw new Error("Unexpected page value " + JSON.stringify(value));
-
-			dateFrom = resolvedQuery.dateFrom;
-			dateTo   = resolvedQuery.dateTo || new db.Date();
-			mode     = resolvedQuery.mode;
-			durationInTimeUnits     = calculateDurationByMode(dateFrom, dateTo, mode);
-			resolvedQuery.pageCount = Math.ceil(durationInTimeUnits / itemsPerPage);
-			if (num > resolvedQuery.pageCount) throw new Error("Page value overflow");
-			return num;
-		}
-	}, serviceQuery, certificateQuery, stepStatusQuery);
+	handlerConf.push(pageQuery, serviceQuery, certificateQuery, stepStatusQuery);
 	queryHandler = setupQueryHandler(handlerConf,
 		location, '/flow/by-role/');
 
@@ -226,7 +164,7 @@ exports['statistics-main'] = function () {
 			offset.to -= 1;
 
 			currentDate = copyDbDate(dateFrom);
-			floorToUnit(currentDate, mode);
+			floorToTimeUnit(currentDate, mode);
 			while (timeUnitsCount <= offset.to) {
 				if (timeUnitsCount === offset.from && query.page > 1) {
 					dateFrom = copyDbDate(currentDate);
@@ -248,7 +186,6 @@ exports['statistics-main'] = function () {
 		delete serverQuery.status;
 
 		queryServer(serverQuery).done(function (responseData) {
-			console.log('responseData', responseData);
 			data.value = filterData(responseData, assign(query, { dateFrom: dateFrom, dateTo: dateTo }));
 			pagination.count.value   = query.pageCount;
 			pagination.current.value = query.page;
@@ -298,7 +235,7 @@ exports['statistics-main'] = function () {
 			p({ class: 'submit' }, input({ type: 'submit' }))));
 
 	section(pagination);
-	section({ class: "section-primary" },
+	section({ class: "section-primary statistics-table-scrollable" },
 		data.map(function (result) {
 			return table({ class: 'statistics-table' },
 				thead(
